Master Repositories Created
Changed ID's of  all master entities from long to Integer type

+++ QandAPlatform/src/main/java/com/sudarshan/qaplatform/entities/Company.java
@@ -14,7 +14,7 @@ public class Company {
 	@Id
 	@GeneratedValue(strategy=GenerationType.AUTO)
 	@Column(name = "ID")
-	private long id;
+	private Integer id;
 	
 	@Column(name = "Name")
 	private String name;
@@ -23,17 +23,17 @@ public class Company {
 
 	}
 	
-	public Company(long id, String name) {
+	public Company(Integer id, String name) {
 		super();
 		this.id = id;
 		this.name = name;
 	}
 
-	public long getId() {
+	public Integer getId() {
 		return id;
 	}
 
-	public void setId(long id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
