+++ QandAPlatform/src/main/java/com/sudarshan/qaplatform/entities/SubTopics.java
@@ -20,7 +20,7 @@ public class SubTopics {
 	@Id
 	@GeneratedValue(strategy=GenerationType.AUTO)
 	@Column(name = "ID")
-	private long id;
+	private Integer id;
 	
 	@Column(name = "Name")
 	private String name;
@@ -33,18 +33,18 @@ public class SubTopics {
 	public SubTopics() {
 		
 	}
-	public SubTopics(long id, String name, Topics topics) {
+	public SubTopics(Integer id, String name, Topics topics) {
 		super();
 		this.id = id;
 		this.name = name;
 		this.topics = topics;
 	}
 
-	public long getId() {
+	public Integer getId() {
 		return id;
 	}
 
-	public void setId(long id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
