+++ QandAPlatform/src/main/java/com/sudarshan/qaplatform/entities/Topics.java
@@ -15,7 +15,7 @@ public class Topics {
 	@Id
 	@GeneratedValue(strategy = GenerationType.AUTO)
 	@Column(name = "ID")
-	private long id;
+	private Integer id;
 	
 	@Column(name = "Name")
 	private String name;
@@ -24,17 +24,17 @@ public class Topics {
 		super();
 	}
 
-	public Topics(long id, String name) {
+	public Topics(Integer id, String name) {
 		super();
 		this.id = id;
 		this.name = name;
 	}
 
-	public long getId() {
+	public Integer getId() {
 		return id;
 	}
 
-	public void setId(long id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
